backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.15.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
boto3==1.34.141
botocore==1.34.141
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.1
crispy-bootstrap5==2024.2
cryptography==42.0.5
cycler==0.12.1
databases==0.9.0
DateTime==5.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dj-database-url==2.2.0
dj-email-url==1.0.6
dj-rest-auth==6.0.0
Django==5.0
django-allauth==0.61.1
django-auto-logout==0.5.1
django-axes==6.5.1
django-cache-url==3.4.5
django-cors-headers==4.3.1
django-crispy-forms==2.2
django-debug-toolbar==4.3.0
django-dotenv==1.4.2
django-environ==0.11.2
django-formtools==2.5.1
django-graphql-jwt==0.4.0
django-ipware==7.0.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-recaptcha==4.0.0
django-storages==1.14.3
django-twilio==0.14.3.3
django-two-factor-auth==1.16.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-spectacular==0.27.2
email_validator==2.2.0
environs==11.0.0
etils==1.7.0
executing==2.0.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.48.1
gast==0.5.4
graphene==3.3
graphene-django==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
hiredis==2.4.0
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
idna==3.7
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.2
ipython==8.21.0
jax==0.4.26
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.1
more-itertools==10.3.0
msgpack==1.0.8
mypy-extensions==1.0.0
oauthlib==3.2.2
openai==1.30.1
passlib==1.7.4
pathlib==1.0.1
pathspec==0.12.1
phonenumbers==8.13.37
pillow
platformdirs==4.2.0
pluggy==1.4.0
promise==2.3
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.6
pure-eval==0.2.2
pvporcupine==3.0.2
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pymupdf
pyparsing==3.1.1
PyPDF2==3.0.1
pypng==0.20220715.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-ipware==3.0.0
python-multipart==0.0.9
python-slugify==8.0.4
python-ulid==1.1.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.7
redis-om==0.3.1
referencing==0.35.1
repath==0.9.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9
s3transfer==0.10.2
sendgrid==6.11.0
sentry-sdk==2.11.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlmodel==0.0.21
sqlparse==0.4.4
srt==3.5.3
starkbank-ecdsa==2.2.0
style==1.1.0
termcolor==2.4.0
text-unidecode==1.3
text2numde==1.0
threadpoolctl==3.3.0
tinydb==4.8.0
toml==0.10.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
twilio==9.1.0
typer==0.12.3
types-cffi==1.16.0.20240331
types-pyOpenSSL==24.1.0.20240722
types-python-dateutil==2.8.19.20240106
types-redis==4.6.0.20240726
types-setuptools==71.1.0.20240726
typing_extensions==4.11.0
tzdata==2024.1
update==0.0.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1